+++ jobqueue/static/js/wr/modal-handlers.js
@@ -38,6 +38,7 @@ const modifyFormObservableFields = [
     'withDocker',
     'withSingularity',
     'containerMounts',
+    'containerImageUser',
     'env',
 ];